backend/requirements.txt
# GovAI 后端依赖
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
redis==5.2.0
python-multipart==0.0.22
httpx==0.27.0
alembic==1.13.0

# 文档处理（降级备选，主引擎在 converter 微服务）
python-docx==1.1.0
json_repair>=0.30.0

# 高精度 PDF 导出（Playwright Chromium 无头渲染）
playwright>=1.40.0
jinja2>=3.1.0
